+++ include/modules/base/RArm.h
@@ -1,4 +1,4 @@
-/*
+ /*
 Copyright (c) 2017 "University of Denver"
 This file is part of Pineapple.
 Pineapple is free software: you can redistribute it and/or modify
@@ -27,9 +27,9 @@ class RArm :
 public:
 	RArm();
 	~RArm();
-	void set_shoulder(double roll, double pitch);
-	void set_elbow(double roll, double yaw);
-	void set_wrist(double yaw);
+	void set_shoulder(float roll, float pitch);
+	void set_elbow(float roll, float yaw);
+	void set_wrist(float yaw);
 	std::string test_message();
 	//Bounds...there's got to be a lazier way to do this...
